+++ CHANGELOG.md
@@ -7,3 +7,10 @@ Review doc in github
 ## [1.0.0]. 25/01/2026   
 Add CHANGELOG.   
 
+## [1.0.1] 26/01/2026
+Change *ignoreComments* for a double choice : 
+    search:
+    code: on    
+    comments: on 
+That allows to search only in comments, sometime usefull. 
+Docs were adapted.